Include full version in pkg-config files
This commit makes a best-effort attempt to append the full
git-describe output to the Version field of generated `.pc`
(pkg-confg) files:
`Version: 2020.10.3 (gex-2020.10.3-32-g0936c7d)`

In a git clone this information is obtained by running `git`, while in
a generated tarball (including releases, plus snapshots such from
`stable`) this is obtained from the `version.git` in the source
directory.  The `version.git` file's content is also used if the
`.git` directory exists, but `git describe ...` fails.

As a special case, if the describe output matches the pattern
`/^gex-$(VERSION)-0-/`, then this is taken to indicate a release, and
no parenthetical is appended to the `Version: ` line.