ibv: fix bug 4141: SRQ-specific test failures
As described in detail in Bug 4141, we have been erroneously
under-sizing the Send scatter-gather list of the Request-only QP
(which is specific to SRQ).

This commit resolves the observed problem by ensuring that the
scatter-gather list contains enough entires for the most aggressive
use of local firehose (determined by `GASNETC_SND_SG`) in the
sending of an AM Long payload.  This is also the length needed
for all other QPs, where a Put may exercise this length.

Since two-entry gather-on-send is sometimes used to inject AM Medium
and packed-Long messages, this commit additionally imposes a floor of
two entries, guarding against the possibility of `GASNETC_SND_SG=1`.

Finally, since there is now only one possible value for
`max_send_sge`, this commit removes an unnecessary argument to
`gasnetc_check_inline_limit()`.