ibv: fix Bug 4173 - race in firehose on powerpc
This commit resolves bug 4173.  As described in more detail in the bug
report, the "interlock" between the injecting thread and a polling
thread running a firehose callback was missing an acquire fence.  This
lack was observed on x86 / x86_64 CPUs where the code was originally
developed, because there every atomic operation includes a *full*
memory fence.