ibv: fix firehose vs CommitRequestLong conflict
This commit resolves Bug 4174:
"testcore2 broken by CHECK_INJECT with LARGE and EVERYTHING".

Instead of "closing" the Prepare/Commit interval (via resetting the
`_magic` to the "BAD" state) as the final step in the reference
implementation of `gex_CommitRequestLong()`, this is made the *first*
step.  This is sufficient to allow the implementation of RequestLong
to inject additional communication such as is needed by ibv-conduit
for a Long destination address missing in the remote firehose cache
(LARGE and EVERYTHING segment modes).

In the future this ability to communicate in Commit may be needed to
manage cached/lazy-learned segment information for RDMA of Long
payloads.  So eventually this may not be a firehose-specific concern.

To retain symmetry for ease of maintenance, the same change is made to
the other three reference Commit implementations.

Nothing has been done to modify the native NPAM Medium implementations
in aries- or ibv-conduit to use the new "consume" call.  While doing
so would be correct, it is not currently necessary and may introduce
additional risks as documented in comments.