Relocate per-thread medium buffer allocation
In anticipation of using it for NPAM, this commit relocates the logic
for per-thread medium buffer (revised in the previous commit) to
earlier in `gasnet_am.h`, bringing with it the prerequisite
definitions of `GASNETC_MAX_{ARGS,MEDIUM,LONG}_NBRHD`.

This commit is *only* movement - no changes to code or comments.