gasnet_attach: promote 0-byte segment to page size
This commit modifies the implementation of `gasnet_attach()` to
allocate a one-page segment when zero-length is requested.

This is motivated by its ability to "close a gap" between the legacy
"attach" API and the GASNet-EX specification for bound segment queries
where a zero length is used to indicate no segment is bound.

This change is *not* entirely transparent to legacy GASNet-1 clients.
For instance, the output of `gasnet_getSegmentInfo()` will show the
non-zero size.  It is also conceivable that allocating a segment where
none was allocated previously may lead to resource problems (for
instance, one additional SysV shared memory region per process).
However, the likelihood of these issues occurring in practice has been
judged to be acceptably low.