Provide thread-safety in handler registration
This commit resolves bug 4175
  "thread-safety of gex_EP_RegisterHandlers()"
by providing both the code and documentation changes requested there.

The required thread-safety is provided by a new per-EP mutex, which
is acquired and released in `gasneti_amregister_client()`.  This is
sufficient to cover the client-facing `gex_EP_RegisterHandlers()`
and `gasnet_attach()` calls.

As noted in new comments, the new serialization does not cover the
internal manipulations of the handler table within `gex_Client_Init()`
and `gex_EP_Create()`.  None is necessary for those paths, since the
client cannot possibly make concurrent calls on the same EP, and
the "noise" of adding it to internal paths didn't seem justified for
no perceived benefit.