fixup: NPAM: (mostly) eliminate malloc/free fro...
This commit corrects an error in the revision of loopback (same
process) handling of NPAM Long with GASNet-allocated buffers.  While
the prepare path had been updated to use new allocation logic, the
commit logic had not received corresponding changes to the free logic.

The logic for intra-nbrhd "IMMEDIATE failure" was also lacking the
corresponding updates.  However, it is not possible to reach that path
with a buffer allocated and thus the out-of-date free calls were
unreachable.  Both instances (Request and Reply) have been replaced by
an assertion that there is no buffer to free.