tests: improve handling of large ERROR outputs
This commit imposes a per-test limit (default 16) on the number of
"ERROR" lines to be included in the final summary output of `make
run-tests` and similar.

Not only does this provide for a more readable summary, but it also
avoids crashing the shell as seen with massive (2GB+) error outputs.