kinds: per-conduit compile/link of kinds objs/libs
This commit revises handling of kinds-specific "special objs" and
values for LDFLAGS and LIBS to be per-conduit "opt-in".  The net
result is that in a build with the CUDA_UVA kind enabled only
ibv-conduit will (a) compile `gasnet_cuda_uva.c` into its
`libgasnet-*.a` and (b) include `-lcuda` (and matching `-L...`) in the
linker flags.

A conduit's `Makefile.am` now lists its supported kinds in a new
`$(CONDUIT_KINDS)` variable, while the `conduit.mak.in` adds the
kinds-specific linker flags to `$(CONDUIT_{LIBS,LDFLAGS})`.  In all
cases, "upstream" logic in `configure` or `Makefile-conduit.mak.in`
ensures that the corresponding substitutions or make variables are
empty for kinds not enabled at configure time (relieving the conduit
Makefile of any responsibility for conditional logic).

This approach additionally recognizes that in a given build different
conduits could support different subsets of the configure-enabled kinds.
In particular, there are no longer any `GASNET_KINDS_{LIBS,LDFLAGS)`
substitutions for the concatenated linker flags of all enabled kinds.