testgasnet: check for graceful failures
This commit adds a section to `testgasnet` specifically for testing of
"graceful failure" of features which are missing/incomplete such as
due to the incremental development on memory kinds.  These are run
with verbose error reporting disabled (motivating the grouping).

Currently the only interface tested here is `gex_EP_Create()` for
conduits with `GASNET_MAXEPS == 1`.