ibv: NPAM Long via RDMA
This commit implements a RDMA protocol for GASNet-allocated NPAM Long,
where previously only a "packed" protocol was available.  The new
environment variable `GASNET_PACKEDLONG_ALLOC_LIMIT` is added to
control the protocol switch point (default is to match the value of
`GASNET_PACKEDLONG_LIMIT`).

Relative to the packed protocol, this saves the `memcpy()` at the
target at the expense of adding a second ibv-level message injection
at the initiator.