Remove all remaining Itanium/IA-64 support
This commit removes all logic related to support of the IA-64
architecture, such as inline asm and compiler intrinsics used for
atomics, timers and memory fences.  Additionally, comments referencing
implementation of (or plans for) IA-64-specific support have been
removed.

As noted in bug 4190, we have lacked access to the IA-64 architecture
for some time, and even Intel is finally planning to cease shipments
later this year.