firehose: warn on oversize RS setting
This commit makes three changes to logic for validating the value of
`GASNET_FIREHOSE_MAXREGION_SIZE`.

The first change is to make the check conditional on the "packed"
format of `fh_key_t`, used when `GASNET_MAXNODES <= GASNET_PAGESIZE`.
This is the case for ibv-conduit (with `GASNET_MAXNODES=65535`) on
PPC64 (with 64K pages), such as Summit.  In the non-packed case, a
full `uintptr_t` is dedicated to the corresponding field, and no
limit is necessary.

The second change is to correct the logic to include a necessary cast,
without which sign-extension was leading to checks against an
impossibly large value.  As a result, a value which could allow
unrepresentable registration lengths was previously permitted.

The third change is to warn the user in the case their value is too
large.