ibv: remove barrier preceding quiesce
As described in bug 4166, it is believed that the bootstrap (non-IB)
barrier preceding `gasnetc_sndrcv_quiesce()` is unnecessary.
This commit removes it.

Initially, `gasnetc_sndrcv_quiesce()` was part of BLCR
(checkpoint/restart) support, removed in commit 72e2bf41.  There it
was called immediately following a call to `gasnet_barrier()` and
nothing has changed since that would suggest that a non-IB barrier is
required.  In fact, the quiesce can actively inject new communication
to recover outstanding credits, making it highly unlikely that it
would be necessary to drain the sends issued by the AM-based reduction
which runs prior to the quiesce call.  That preceding reduce-to-all is
sufficient to provide a barrier synchronization *if* such is required
(it is probably not).

NOTE:
The quiesce logic was used for checkpoint/restart in the past and is
anticipated to be part of a possible `gex_Client_Fini()` in the
future.  However, it serves no necessary purpose presently and so is
compiled and run only when `GASNETI_IBV_SHUTDOWN` is defined (such as
via `MANUAL_CFLAGS`).  This allows testing to help avoid "bit rot".
In particular, as new IB-level resource allocation is added for
multiple endpoints and segments, the shutdown will fail (at
`ibv_dealloc_pd()` if not before) if the balancing deallocations are
not added.