ofi: fix bug 3747 (link error w/ generic atomics)
This commit resolves

  * Bug 3747 - ofi-conduit PAR link failure with generic atomics

by adding a configure-time check for native (or compiler) atomics.
For the platforms where libfabric is supported, this is equivalent to
the desired check for spinlock support.  So, configure now refuses to
build ofi conduit with generic atomics.  The same was already true of
ibv for different reasons.