ibv: repair --enable-stats builds
This commit corrects preprocessor logic added in commit b15eac43 to
prevent crashes in debug builds.  That change identified tracing as
the subset of debug features contributing to the problem.   However,
at least currently, stats-only builds suffer from the same issue.

It has not been determined if the original commit was incorrect, or if
the problem for a stats-only build arose due to later changes.  If the
former, then this bug has been unreported since the 2018.3.0 release.