ssh-spawner: resolve bug 4203 w/ a better message
This commit resolves

 * Bug 4203 - Improve error message when passwordless ssh isn't set up

by describing the likely problem, including the host the connection is
originating from.  It is worth noting that one might also reach this
message if the ssh was OK but a firewall blocked the child-to-parent
connection.  However, that will be distinguishable in verbose spawner
output by the presence of error messages about the `connect()`
failure.