Add K/F for bug 4207
This commit adds the narrowest practical K/F for bug 4207 as it
as been initially reported.  Only the (procs!=1) aspect of the
original report is not reflected.

As more Sun C testers run, this K/F may be added to the -seq test
and/or have its conditional expanded.  But for now, we want to see any
such additional failures rather than hide them as KNOWN.