ibv: resolve bug 4196 dynamic connect + multi-ep
This commit resolves bug 4196 "ibv: dynamic connect does not support
multiple endpoints" by removing the assertion which had been seen to
fire.  As describe below, it has been determined to be unnecessary.

A careful review of the dynamic connect code shows that due to the
current *incomplete* implementation of multi-ep in ibv-conduit, things
"just work" to complete the correct connection.  This is true because
the non-primordial EP's current share the QPs of the primordial ones.
To be more accurate: it is because the primordial and non-primordial
endpoints connected to the same remote rank share the same
`gasnetc_cep_t`, which is the object the connection logic operates on.
This even means that concurrent attempts from different threads to
connect a primordial and non-primordial endpoint to the same peer will
"just work" as well.

Future work to provide distinct resources (isolation) for
non-primordial endpoints will necessarily provide each with a distinct
`gasnetc_cep_t`.  If the connect code is not updated to distinguish
among distinct endpoint indices at that point, then the dynamic
connection code will become erroneous.  However, the removed assertion
would still be incorrect.  So, it is removed rather than commented out
or otherwise temporarily disabled.