ibv: honor GASNET_USE_FENCED_PUTS w/ single HCA
This commit removes warnings about (and discarding of) a "true" value
for `GASNET_USE_FENCED_PUTS` in a build in which multi-rail support
is enabled, but only a single HCA has been enabled/detected.

A non-trivial share of the changes in this commit are to ensure that
the fenced-puts logic is compiled for the case of a build which has
been configured using `--with-ibv-max-hcas=1`, which was previously
considered to be equivalent to disabling multi-rail support.

Why?  We have determined that the issue described in bug 3447 can
theoretically occur with multiple processes per host which taken
together use more than one HCA.  We still desire to compile-away
support for fenced puts in a truly single-rail environment, but now
wish to support it for the case of "single-rail run in a multi-rail
build".  Thus `--with-ibv-max-hcas=1` continues to allow some things
to compile away, but enables the logic for fenced puts.

Documentation has been updated accordingly, including a recommendation
to configure multi-rail (possibly using `--with-ibv-max-hcas=1`) if
one could ever use multiple HCAs per host.