ibv: fix bug 4208 - multirail/pshm/xrc conflicts
This commit revises the XRC logic to allow for the possibility that
not every process in the PSHM supernode (aka GEX nbrhd) is using the
same HCA.  This then allows a "default" `gasneti_nodemapInit()` call
like most other conduit.  The net result is that one can have a
full-host-width supernode, XRC and heterogeneous multirail at the same
time, where previously one could, at best, have two of those three.

As alluded to above, the majority of the changes are focused on a
revision of XRC logic.  Previously the IBV concept of an XRC domain
was taken to coincide with a PSHM supernode.  However, as noted in bug
4208, that assumption is broken when different processes in the same
supernode are using different GASNET_IBV_PORTS values (either
explicitly or via the hwloc-based mechanism).  The use of the HCA's
local identifier (lid) as a host identifier when constructing the
nodemap had hidden the problem by unintentionally sub-dividing a host
along HCA-use boundaries.  However, as the bug report notes, that
has a very bad impact on intra-host communication which would
otherwise be using shared memory.

Fortunately, the lid table and nodemap are sufficient to "discover"
the proper XRC domain boundaries prior to initialization of PSHM and
without added communication.  The ability to perform this discovery
prior to initialization of PSHM is critical to enable allocation of
the correct amount of shared memory for use by the XRC subsystem to
accommodate its (preexisting) sharing of XRC QP numbers and a new
table used to (sometimes) translate ranks to XRC domains.

The *net* shared memory used for storage of XRC QP numbers is
unchanged relative to the situation prior to this commit.  Using the
bug report's example of an HCA per socket on Summit: where a compute
node previously had two supernodes each with a single table, we now
have single supernode with two tables (2*1 == 1*2).  Put another way:
both before and after this commit we have a table per HCA, only now
they are in the same shared address space rather than in two disjoint
ones.  Similarly, the HCA resources used have not changed.

As mentioned above, there is one new data table "sometimes" required
for mapping of ranks to XRC domains.  Prior to this work the 1-to-1
correspondence between XRC domains and supernodes meant that the array
backing `gasneti_node2supernode()` was sufficient.  For jobs where all
processes on a host use the same HCAs (including any job on a
single-rail system) that correspondence is still present and this new
table is not constructed, nor is shared memory allocated for it.
Instead we continue to use `gasneti_node2supernode()` as before.

When the new table _is_ required, it consumes just two bytes per rank
in the job.  Additionally, its contents are constant such that we can
keep a single instance per-supernode in shared memory, rather than
replicating it in every process.  In the future, the two-byte datum
for a given remote rank will likely be among the things cached lazily.

This commit makes changes only to initialization and connection paths.
Excepting the extent to which dynamic connections might be considered
"critical", there are no instructions added to critical paths.