fixup: testcore2: add overwrite testing
This commit corrects logic for overwrite of the source buffer of an
AMReplyLong to *not* overwrite an in-segment source buffer (used in
some fraction of the iterations).  This is necessary because as soon
as the Reply is delivered, the in-segment memory may become the
destination of a new AMRequestLong.

The fix includes both augmenting a conditional to exclude the INSEG
case, and correcting the destination argument to `memset()` which
was incorrectly always overwriting the in-segment payload buffer,
even when the Reply source was a distinct (dynamically allocated)
temporary buffer.  Thus, both the in-segment and out-of-segment
source cases were incorrect prior to this fix.