Don't mark tools-only builds as "dirty"
This commit modifies the build system logic to not pass `--dirty` to
`git describe` commands reachable in a tools-only build.  This is
because a tools-only build from a git clone will always be "dirty"
(due to the action of `Bootstrap -o`) making this marking unhelpful
and even misleading when it appears in distribution tarballs and
resulting installs.