Bootstrap: fix xform to/from tools-only mode
This commit drops "$srdir/" prefixes from file and directory names
used in the transformations to and from tools-only mode.  Prior to
any use of these paths, we have already executed `cd $srcdir`, which
makes the prefixed paths not only unnecessary, but also _incorrect_
for the case that `Bootstrap` was invoked using a relative path.