Ordering of CHECK_INJECT after TRACE
This commit revises the placement of `GASNETI_CHECK_INJECT()` calls to
order nearly all of the checks to follow after tracing.  There are two
intentional exceptions.

The first exception is for Test/Try/Wait calls which trace their
*result*, and thus place tracing after the "meat" of the call (too
late for `GASNETI_CHECK_INJECT()` to precede "bad things).

The second exception is for zero-length RMA operations which no longer
call `GASNETI_CHECK_INJECT()` at all.  A commented-out call is left in
the zero-length check macros to indicate this omission is intentional.