Revise and rename gasneti_pshm_jobrank_if_eligible
This commit makes major changes to `gasneti_pshm_jobrank_if_eligible()`.

The first is that it is no longer compiled conditional on PSHM and
thus has been given a new name: `gasneti_jobrank_if_mappable()` (and
the few callers have been updated accordingly).

The second is that it has been stream-lined for the (still common)
case of builds without any device kinds support.

The third is that it now includes in in-nbrhd check (or is-local when
PSHM is disabled), rather than requiring the caller to check.

Finally, the logic has been added to deal correctly with loopback
access to non-primordial host memory segments.