Implement gasnete_mapped_at()
This commit implements `gasnete_mapped_at()` which is to be deployed
as a replacement for `gasnete_islocal()`.  This new helper uses
`gasneti_jobrank_if_mappable()` to replace the function of
`gasnete_islocal()` as the "can I memcpy() to/from it" query, thus
extending the query to _include_ cross-mapped segments (when PSHM is
enabled) and _exclude_ local device segments.  In addition to
performing the Boolean check, this function returns the locally mapped
address on success (and NULL on failure).