ibv: set MLX[45]_SINGLE_THREADED when possible
This commit resolves Bug 4053:
  ibv: Default to setting MLX5_SINGLE_THREADED=1 for SEQ builds w/o progress thread

For the default settings (configure and environment), a SEQ build of
ibv-conduit does NOT start either the RCV or CONN threads, though
support for both is compiled-in by default.  This commit adds logic to
determine early in initialization if either of those threads will be
started.  It they will not, then `MLX4_SINGLE_THREADED=1` and
`MLX5_SINGLE_THREADED=1` are added to the environment (unless already
set) prior to initialization of libibverbs.

There is an uncommon corner case in which a user requests static
connection to a subset (via `GASNET_CONNECTFILE_IN`) which later turns
out to be fully-connected.  In this case, we will wrongly assume the
connection thread will be started.  Since the connection files are
per-process, communication is required to discover this corner case.
So, there is no plan to resolve it.

Both `testsmall` and `testam` show measurable improvements.