ibv: fix compile-time disable of dynamic connect
This commit repairs compile-time disable of dynamic connect support in
ibv-conduit, such as via `MANUAL_DEFINES=-DGASNETC_DYNAMIC_CONNECT=0`.
Specifically, definition of `conn_get_srq_num()` is moved outside of
`#if GASNETC_DYNAMIC_CONNECT`, since it is needed by XRC support
independent of dynamic connection support.

It is worth noting that this disable exists as an undocumented means
to disable the dynamic connection logic if it is ever suspected of
being the cause of some erroneous behavior.