Implement --enable-(pkg-)rpath options (bug 3420)
This commit implements new configure options `--enable-FOO-rpath`
and `--enable-rpath`, where `FOO` is the name of a conduit or of a
library such as `hwloc` or `pmi`.

By default, no RPATH options are added to any `*_LDFLAGS` variables.
Use of `--enable-rpath` changes that such that RPATH options will be
added for every package by default (if appropriate, see below).

Using `--(en|dis)able-FOO-rpath` provides per-package control,
overriding the default established by `--(en|dis)able-rpath`.  This is
implemented in `GASNET_PACKAGE_LDFLAGS()` defaulting to the same
directory as identified for the `-L` option.  However, since `-L` is
for link time while an RPATH is used at run time, one may need a
different value.  Therefore it is possible to specify an alternate
directory to be used (e.g. `--enable-pmi-rpath=/cluster/slurm/lib`).

As with the `-L` and `-I` options, care is taken not to add an RPATH
entry for the system directories `{,/usr}/lib{,32,64}`.  Additionally,
no RPATH is added to a user-specified `*_LDFLAGS` - only the default
values are changes.  These two conditions are what the phrase "if
appropriate" meant when used above.

Due to the exclusion of standard locations, for systems with all the
needed libraries in those locations there is no change in link flags
when using the new configure options.

Adding of RPATH options to a given `*_LDFLAGS` is "best effort", with
a warning if the proper compiler options could not be identified by
configure probe *and* there was "appropriate" occasion to use them.