testam: make warm-up conditional
Prior to this commit, all warm-ups iterations in `testam` were
performed unconditionally.  This commit provides for both coarse- and
fine-grained conditionals.

For the coarse-grain, this commit provides a `-s` command line option
to skip warm-up iterations entirely.

For the fine-grain, this commit ensures that warm-ups for Short, Medium
or Long are performed only when at least one corresponding section is
to be run.  To enable this, a new `TEST_SECTION_ENABLED_FWD()` has
been added to `test.h` to allow "looking ahead" for enabled sections.

Additionally, this commit bounds the number of warm-up iterations at
10,000.