Remove dynamic auxseg checks from gex_RMA paths
This commit removes dynamic `gasneti_in_auxsegment()` range checks
from the critical paths of shared-memory RMA operations.  Instead,
the code now asserts that the remote address is not in the remote
aux seg.

The choice to assert is based on a determination that no current code
performs Put or Get to/from an in-nbrhd aux segment.  The nearest case
is RMA-based dissemination barriers, which perform their own reduction
to load/store.  Should this change, the logic can be updated to use
the same `GASNETI_FLAG_PEER_SEG_AUX` flag deployed for intra-nbrhd
Longs.