ssh-spawner: fix bug 4230 (flawed de-duplication)
This commit resolves
  Bug 4230 - ssh-spawer de-duplication logic is flawed

The logic to de-duplicate the list of hosts was previously applied to
a list which had already been capped at NNODES entries, meaning that
many cases (including the per-core replication of each hostname) would
not contain entries which should have become relevant post-de-dup.

This commit moves the de-duplication "inline" into the parsing of node
lists from an environment variable or stdio stream.  The result is
that when de-duplication is enabled (the default) processing of
hostnames terminates when at most NNODES *unique* hostnames have been
processed.  This resolves the originally reported issue.

Removal of *redundant* de-duplication logic in the parsing of a stream
resolves the second issue in the bug report (comment #2).  In addition
to creating a difference in handling `GASNET_SSH_NODEFILE` and
`GASNET_SSH_SERVERS` with identical lists, this defect could have
defeated the use of `GASNET_SSH_KEEPDUP=1` for fine-grained control
over job layout.

In addition to fixing the two defects in the bug report, some related
cleanups have been applied:

1) Use `gasneti_getline()` to avoid an arbitrary limit on the length
of a line read from a stream.

2) Restructure parsing logic to avoid needless strdup+free of
hostnames which are rejected as duplicates.

3) Use of C99 mid-block declarations and declarations in for loop
headers.