Improve help for --enable-FOO-rpath
This commit updates the logic generating `configure --help` output
for the `--enable-FOO-rpath` options to omit the output when the
help output for the corresponding `FOO_LDFLAGS` has been suppressed.

This eliminates output for `--(en|dis)able-pmi-rpath` from among the
options relating to aries-conduit configuration, allowing it to
instead appear together with the `--with-pmi-ldflags` help.