aries: resolve bug 4233 - multi-domain FPE at init
This commit resolves Bug 4233 "Multi-domain regression in
aries-conduit leads to SIGFPE during init" by performing the
initialization of multi-domain in `gasnetc_init()` prior to the call
to `gex_EP_Create()`.

As noted in the bug report, the problem arises from a call to
`GASNETI_CHECK_INJECT()` added to `gex_EP_Create()` in pull request
452, which leads to DEBUG builds calling `gasnete_init_threaddata()`
for the initial thread far earlier than before.  In particular, the
threaddata for the thread calling `gex_Client_init()` is initialized
before `gasnetc_threads_per_domain` is set, resulting in division by
zero.

This commit resolves the problem by splitting the logic to initialize
some key state for multi-domain from the logic to initialize
messaging.

The alternative approach of reordering the call to `gex_EP_Create()`
to *after* `gasnetc_init_messaging()` was rejected because it would
require introducing an additional barrier prior to the call to
`gasnetc_sys_coll_init()` to ensure EPs are globally ready to receive
AMs.  A comment has been added to document this (previously unstated)
requirement.