Annotating fixed-before-nvhpc work-arounds
This commit notes several PGI-specific bug work-arounds which appear
NOT to be applicable to NVIDIA's "NVHPC" branded compilers.

The use of `gasneti_empty_function()` in the `GASNETI_UNUSED_ARG*()`
macros (`gasnet_basic.h`), and the tweak to `_GASNETI_THREAD_POSTED`
have been confirmed unnecessary with the NVHPC compilers, after
successfully reproducing the reported issues with older PGI compilers.

The remaining work-arounds related to tracing are not reproducible
with PGI compilers available to me, let alone the current NVHPC
releases.

There are NO warnings from `test{gasnet,legacy}.c`.

Testing was conducted with `-w` removed from the compiler flags, and
with appropriate configure options to ensure the work-arounds are
reachable.