configure: generalize GASNET_CRAY_POST_LINK_OPTS
This commit generalizes the `GASNET_CRAY_POST_LINK_OPTS` m4 macro and
renames it to `GASNET_CRAY_FIX_LINK_OPTS`.  The generalizations are
(1) to allow the caller to specify a different variable to "fix", and
(2) allow for the directory to occur multiple times in the variable's
value (such as for `-L[dir]` and `-Wl,rpath,[dir]`).