spawner: try harder when failing to ID executable
This commit modifies the (already non-fatal) warning path taken when
the `gasnetrun_[conduit].pl` script fails to find an executable which
has the expected ident strings.  Where previously the script would
issue a warning and then run with the *last* word of the command line
as the executable, the logic now adjusts state as necessary to treat
the *entire* command line (after the script's own options) as the
GASNet application.  This is consistent with the apparent intent of
having a warning (not an error) and can work for the cases where the
GASNet executable is run from a script, or in which the GASNet
portions are (ick) dynamically loaded such that this script cannot
be expected to see them.