configure: avoid unused -L and rpath in LDFLAGS
This commit makes an improvement to the logic to generate per-package
LDFLAGS.  Prior to this commit, if the named library could not be
found under the package's prefix, then `-L$prefix/lib` would be
used for the LDFLAGS, even though the desired library was not in
that directory (which might not even exist).  Similarly, there
could be useless RPATH options generated.  Now one gets an empty
per-package LDFLAGS if no directory was found to contain any of
the requested libraries.

The scenario described can occur in practice when the headers for a
given package are in `$prefix/include` but the libraries are installed
elsewhere, such as within the system's standard library search path.