testtools: eliminate undefined underflow/overflow
This commit replaces use of signed `gasnett_atomic_sval_t` with its
unsigned sibling `gasnett_atomic_val_t` for a temporary variable
used in testing of wrap-around properties of `gasnett_atomic_val_t`.

The previous code computed `GASNETT_ATOMIC_SIGNED_MIN - 1` and
`GASNETT_ATOMIC_SIGNED_MAX + 1` using the signed type.  However, this
is undefined in both C and C++.  By using the unsigned temporary the
behavior is well defined (matching documentation in README-tools).
This does not "damage" the test which is checking a property of the
atomic type and the constant definitions, not the properties of the
`gasnett_atomic_sval_t` type.