ibv: reduce space required by gasnetc_cep_t
This commit resolves bug 4240 "ibv: wasteful layout of gasnetc_cep_t"
by reorganizing the in-memory layout of ibv-conduit's `gasnetc_cep_t`.
Changes include:

+ Correct the grouping of read/write and read-only fields
+ Exactly one cache pad at the end of each of those groups
+ Fields replicated from cep->hca are noted in comments
+ Each group is sorted by size to get dense (padding-free) layout,
  independent of the ABI and choice of optional features.
+ Final padding for cache alignment is sized accurately to achieve a
  total size which is a multiple of the cache line size.

On Dirac (64-byte cache line) and Summit (128-byte cache line) with
multi-rail, XRC and SRQ all enabled:
```
          Dirac      Summit
Before:   288 bytes  488 bytes
After:    192 bytes  256 bytes
```

Currently the only known layout inefficiency is the size of the cache
padding between the R/W and R/O groups.  Because `gasneti_semaphore_t`
contains its own padding to cache line size, it is placed last in the
R/W group and serves as the separation to ensure the two groups are
never on the same cache line.  However, this means that the initial
(currently eight) bytes of the first R/O cache line are unused.
However, this is the best we can do without either breaking the
abstraction of the semaphore or separating the R/W and R/O groups into
distinct data structures.  The latter is a realistic option which
would save *more* than eight bytes because a struct entirely of R/O
fields would not need any cache padding.  However, that is beyond
the scope of the current work.