Implement GASNET_BACKTRACE_MT (bug 4237)
This commit implements environment variable `GASNET_BACKTRACE_MT` to
force multi-threaded or single-threaded backtraces (when the given
debugger has MT support), overriding the default (which follows
the value of `GASNETI_THREADS`).

Care has been taken to avoid introducing dynamic memory allocation
or calls to `strlen()`, making the code marginally less clear than
the most direct implementation in favor of being robust if generating
a backtrace in response to memory corruption.

This has been tested for `gdb`, `idb` and `pgdb` using `testexit`
built in both PAR and SEQ modes, with test numbers 100 and 300.  All
three debuggers produce sufficiently distinct output for single- and
multi-threaded modes that it was simple to verify that the behavior
followed `GASNET_BACKTRACE_MT` and that the default was unchanged.

There are no other debuggers for which we have distinct commands for
single- and multi-threaded backtraces.

This resolves bug 4237: "Option for backtrace to provide the stack on
all threads even in threadmode=seq"