ibv: tweak snd_cq_sema_p references
This commit follows up on observations made in the course of work on
PR#465 (to fix layout issues in `gasnetc_cep_t`) related to the
`snd_cq_sema_p` fields in both that struct and in `gasnetc_hca_t`.

1. One use in a critical communication path (retiring CQ entries) has
been simplified to remove a pointer indirection, based on the
observation that the `hca` is already a "live" variable at this point.

2. Initialization (not performance-critical) of the field in
`gasnetc_cep_t` has been rewritten mainly for improved clarity, but
also removes unnecessary instructions.