Add static qualifier to backtrace string constants
This commit adds the `static` qualifier to existing `const char[]`
declarations used in the backtrace and abortive exit code paths.  This
is intended to ensure that even a dumb compiler will not attempt to
instantiate these string constants on the stack, since this code is
expected to run in response to "bad things" which may include
constrained stack space.