fixup: ibv: reduce space required by gasnetc_cep_t
This commit repairs a static assert to allow for internal padding.
This includes padding at the end of the struct when its "natural" size
is a multiple of 4 but not of 8 (on LP64) as well as possible internal
padding after the semaphore, which is only 4 bytes for `(!THREADS &&
NDEBUG)`.

The total padding is unchanged by this commit, but the intermediate
computations have been revised to not depend on the presence of
cache padding in the sema_t.