Introduce GASNETI_SPIN_*() family of macros
This commit introduces a `GASNETI_SPIN_*()` family of macros to
implement some common patterns which follow the recommended idiom for
use of `GASNETI_WAITHOOK()`.  These differ from the pre-existing
`gasneti_{poll,wait}{while,until}()` in two key ways.  The first is
providing more generality by accepting the loop body as a macro
argument.  The second is the lack of an RMB in the new macros.

Additionally, the pre-existing `gasneti_{poll,wait}{while,until}()`
have be reimplemented in terms of the new macros plus an RMB, and
their comments have been updated to mention the RMB difference.

Finally, the block comment which accompanies the definition of
`GASNETI_WAITHOOK()` has been revised to include appropriate
references to the new macros.

Future work is expected to use these new macros to replace many direct
uses of `GASNETI_WAITHOOK()`, especially ones which currently deviate
from the recommended idiom.