Simplify implementation of GEX_FLAG_AD_{ACQ,REL}
This commit resolves bug 4264:
   simplify implementation of GEX_FLAG_AD_{ACQ,REL}

The "meat" of this commit is the change in gasnet_ratomic.h from an
expression involving two `?:` operations and an `|`, to a single `&`.
The remainder of the commit supports that change by ensuring equality
between the respective `ACQ` and `REL` constants in the `GEX_FLAG_AD_`
and `GASNETI_ATOMIC_` families.

The textually largest change is a change to several `GEX_FLAG_*`
constants.  To enable the desired alignment between constants, the
following changes are made:

+ `GEX_FLAG_AD_{ACQ,REL}` have been moved to the `(1<<1)` and `(1<<2)`
  positions to match `GASNETI_ATOMIC_{ACQ,REL}`.
+ `GEX_FLAG_{SELF,PEER}_SEG` have shifted up by two positions to avoid
  conflicts with the new values of `GEX_FLAG_AD_{ACQ,REL}`.
+ `GEX_FLAG_AD_MY_{RANK,NBRHD}`, `GEX_FLAG_ENABLE_LEAF_LC` and
  `GEX_FLAG_AM_PREPARE_LEAST_ALLOC_{CLIENT,ALLOC}` have also shifted
   up two positions, due to the new positions for the segment
   assertion flags.

Additionally, to avoid ever breaking the newly established equalities,
comments have been added to both sets of definitions and a pair of
static assertions have been added.

The performance impact of these changes was evaluated using a version
of `testfaddperf` modified to pass a variable for `flags`.  The value
was 0 in testing, but dependent on command line arguments to ensure
the compiler could not optimize based on the value.  Additionally, the
number of digits in the latency output was increased to report
picosecond resolution.

On Dirac, means over 40 runs each for pre- and post-commit versions of
`faddperf 500000000 A` show latencies for FADD on U32 dropping 2.5%: from
(14.69 +- 0.03) ns to (14.33 +/- 0.08 ns), with two processes and
PSHM.  For the non-PSHM (single-process) case the reduction is 3.1%:
from (12.45 +/- 0.04) ns to (12.06 +/- 0.03) ns.