smp optimizations for GEX_FLAG_AD_MY_* flags
This commit resolves bug 4263:
        remove "AD_MY_NBRHD" check on smp-conduit

For the case of smp-conduit with PSHM enabled, the reference ratomics
implementation now assumes `GEX_FLAG_AD_MY_NBRHD` and does not check
for this flag bit.  Similarly, smp-conduit without PSHM, the
implementation now assumes `GEX_FLAG_AD_MY_RANK` and does not check
for this flag bit or for `GEX_FLAG_AD_MY_NBRHD`.

The performance impact of these changes was evaluated using a version
of `testfaddperf` modified to pass a variable for `flags`.  The value
was 0 in testing, but dependent on command line arguments to ensure
the compiler could not optimize based on the value.  Additionally, the
number of digits in the latency output was increased to report
picosecond resolution.

On Dirac, means over 40 runs each for pre- and post-commit versions of
`faddperf 500000000 A` show latencies for FADD on U32 dropping 18.5%:
from (14.33 +- 0.08) ns to (11.68 +/- 0.03) ns, with two processes and
PSHM.  For the non-PSHM (single-process) case the reduction was 18.7%:
from (12.06 +/ 0.03) ns to (9.80 +/- 0.03) ns.