aries: update several spinloops
This commit updates several spinloops in gasnet_aries.c to use the new
`GASNETI_SPIN_*()` macros.  In all cases the alternative expression is
more concise (and more expressive, IMO).  In some case, the use of the
new macros introduces the desirable "peeling" of the initial check to
avoid a `GASNETI_WAITHOOK()` in the event the loop terminates after
the first execution of the body.

Additionally, where initial conditionals are manually "peeled" this
commit removes use of branch prediction (`if_pf()` or `if_pt()`),
which we believe to be "premature optimization".