coll/reduce: minor cleanups
This commit removes a redundant declaration and a redundant statement,
both in the implementation of reductions.  These "code changes" are
not "logic changes".

The removed declaration shadowed an identical one in an enclosing
scope.

The removed statement was an assignment right before a `goto` which
is followed by an identical assignment.