ucx: correct configure-time probe for libdir
This commit corrects the configure probe for UCX's libdir.

Previously the probe was looking for `libucx`, which does not exist.
This commit simply corrects one character to look for `libucp` which
we link explicitly.

I was surprised that this has gone unnoticed until now.  However, it
appears likely that all ucx-conduit testing (until today) has been on
platforms where the UCX libdir appears in the linker search path
either by virtue of a system-level install (placing libs in
`/usr/lib64`) or due to an `LD_LIBRARY_PATH` setting by an environment
module.