ibv: implement GASNET_IBV_PORTS_VERBOSE
This commit resolves Bug 4246 "RFE: better behavior of single-rail
build on multi-rail system" by implementing (and documenting) a new
env var `GASNET_IBV_PORTS_VERBOSE`.  In the spirit of env var
`GASNET_ODP_VERBOSE`, a value of 0 suppresses the warning which is the
subject of the bug, while values larger than 1 can be used to provide
additional detail in the future (such as providing a listing like the
one requested by `GASNET_IBV_LIST_PORTS`).