Support root/leaf mix in array-based NB event APIs
As reported in bug 4218, the use of `gex_Event_WaitAll()` on an array
containing both the root event and alc leaf event from
`gex_RMA_PutNB()` has been seen to fail with the message:
```
*** FATAL ERROR (proc 0): in gasnete_test_array() at [redacted]/gasnet_event.c:395: Invalid use of free eop
```
It is believed that all of `gex_{Test,Wait}{Some,All}()` can be made
to fail in this manner when passed an array containing leaf events
and their parents (in the right states).

At present, `testlarge.c` contains the following logic to avoid
triggering the issue:
```
  if (doalc) gex_Event_WaitAll(events+iters, iters, 0); // leaf events first...
  gex_Event_WaitAll(events, iters, 0); // ...then root events
```

Prohibiting the mixing of leaf and root events was *not* our original
intent.  However, the documentation had failed to make our intention
clear and lack of a test case allowed the implementation to
unintentionally depart from the original intent.

The "problem" was a logic error in the debug checking code in
`gasnete_event_check()` generating the message above for both "free"
and "pendingfree" states of the eop.  The pendingfree state was
created specifically for use by the logic in `gasnete_test_array()`
(backing `gex_Event_WaitAll()`) to place root events "in limbo" until
their corresponding leaf events (if any) can be processed as well.
However, as written the `gasnete_event_check()` call on the leaf would
observe the pendingfree state of its root event as erroneous.

The change in this commit is to separate the treatment of "free" and
pendingfree (for both eops and iops) to allow for a leaf event to be
pendingree, while continuing to treat pendingfree root events as
erroneous.

Distinct commits will update the documentation and add test coverage
to `testlarge`.